creating new debugging and working on chart

+++ project/nativeReplacements/String.h
@@ -1,7 +1,8 @@
 #pragma once
 #include <string>
 #include <iterator>
-
+#include <cstdint>
+#include <iostream>
 class String: public std::string {
     public:
     String(const char* str): std::string(str) {};
@@ -12,19 +13,19 @@ class String: public std::string {
     ~String() {};
     bool equals(String& other) { return *this == other;}
     bool equals(const String& other) { return *this == other;}
-    int toInt() {
-        return std::stoi(*this);
+    long toInt() {
+        return std::stol(*this);
     }
     float toFloat() {
         return std::stof(*this);
     }
 
-    size_t write(u_int8_t c) {
+    size_t write(uint8_t c) {
         this->push_back(c);
         return 1;
     }
 
-    size_t write(const u_int8_t* s, size_t n) {
+    size_t write(const uint8_t* s, size_t n) {
         for (size_t i = 0; i < n; i++)
         {
             this->push_back(s[i]);
